Finish first pass at generate_guess
Currently, the function filters the set of possible codes every time we
have a new guess based on the hints on last time. Currently, after the
filter pass, the guess is randomly picked from the remaining codes.

There's probably something we can do here to get a better hit rate, but
it works for the meantime, and we aren't done yet!